+++ OperatorsAndExpressionsHomework/16.BitExchangeAdvanced/BitExchangeAdvanced.cs
@@ -32,6 +32,9 @@ class BitExchangeAdvanced
             Console.WriteLine("Out of range.");
             Environment.Exit(1);
         }
+
+        //My idea is to convert the number to binary stored in a string and
+        //switch parts of it to get the new one and then convert it back to decimal.
         string bitRepresentation = Convert.ToString(number, 2);
         string lowBlock = "";
         string highBlock = "";
@@ -62,6 +65,6 @@ class BitExchangeAdvanced
         {
             exchangedBitNumber += Convert.ToString(bitRepresentation[m]);
         }
-        Console.WriteLine(Convert.ToInt64(exchangedBitNumber, 2));
+        Console.WriteLine(Convert.ToUInt32(exchangedBitNumber, 2));
     }
 }
